# src/main/resources/docs/dgbank-api-docs.yaml
openapi: 3.0.3
info:
  title: DG Bank API
  description: |
    API completa per il sistema bancario DG Bank.
    
    ## Autenticazione
    L'API utilizza JWT Bearer token per l'autenticazione. Dopo il login o la registrazione, includi il token nell'header Authorization:
    `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Token JWT: 24 ore di validità
    - Refresh token: 7 giorni di validità
    
    ## Formati Monetari
    Tutti gli importi sono gestiti con precisione decimale (15,2) per evitare errori di arrotondamento.
  version: 1.0.0
  contact:
    name: DG Bank Development Team
    email: dev@dgbank.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.dgbank.com
    description: Production server
...
paths:
...
  /transactions/deposit:
    post:
      tags:
        - Transactions
      summary: Effettua deposito
      description: |
        Effettua un deposito su un conto (es. da ATM o bonifico esterno).
        Il conto deve essere attivo e l'importo positivo.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
            example:
              accountId: 1
              amount: "500.00"
              description: "Deposito contanti ATM"
      responses:
        '200':
          description: Deposito completato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Errore nel deposito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                inactive_account:
                  summary: Conto non attivo
                  value:
                    success: false
                    message: "Il conto non è attivo"
                invalid_amount:
                  summary: Importo non valido
                  value:
                    success: false
                    message: "L'importo deve essere positivo"
  /transactions/withdraw:
    post:
      tags:
        - Transactions
      summary: Effettua prelievo
      description: |
        Effettua un prelievo da un conto (es. da ATM).
        Controlla che il saldo sia sufficiente.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
            example:
              accountId: 1
              amount: "200.00"
              description: "Prelievo ATM"
      responses:
        '200':
          description: Prelievo completato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Errore nel prelievo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                insufficient_funds:
                  summary: Saldo insufficiente
                  value:
                    success: false
                    message: "Saldo insufficiente"
  /transactions/transfer:
    post:
      tags:
        - Transactions
      summary: Effettua trasferimento
      description: |
        Effettua un trasferimento tra due conti interni della banca.
        Entrambi i conti devono essere attivi e il conto di origine deve avere saldo sufficiente.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
            example:
              fromAccountId: 1
              toAccountId: 2
              amount: "150.00"
              description: "Bonifico a Mario Rossi"
      responses:
        '200':
          description: Trasferimento completato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Errore nel trasferimento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                same_account:
                  summary: Stesso conto
                  value:
                    success: false
                    message: "Non puoi trasferire denaro allo stesso conto"
                account_not_found:
                  summary: Conto non trovato
                  value:
                    success: false
                    message: "Conto di destinazione non trovato"
  /transactions/account/{accountId}:
    get:
      tags:
        - Transactions
      summary: Transazioni del conto
      description: Ottieni tutte le transazioni di un conto (sia in entrata che in uscita)
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lista delle transazioni
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  accountId:
                    type: integer
                    format: int64
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  count:
                    type: integer
                    description: Numero totale di transazioni
  /transactions/account/{accountId}/paginated:
    get:
      tags:
        - Transactions
      summary: Transazioni paginate
      description: Ottieni le transazioni di un conto con paginazione (ordinate per data decrescente)
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          description: Numero della pagina (0-based)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Dimensione della pagina
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Lista paginata delle transazioni
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTransactionResponse'
  /transactions/account/{accountId}/recent:
    get:
      tags:
        - Transactions
      summary: Transazioni recenti
      description: Ottieni le transazioni più recenti di un conto
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          description: Numero massimo di transazioni da restituire
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Lista delle transazioni recenti
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  accountId:
                    type: integer
                    format: int64
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  count:
                    type: integer
  /transactions/{transactionId}:
    get:
      tags:
        - Transactions
      summary: Dettagli transazione
      description: Ottieni i dettagli completi di una transazione specifica
      security:
        - bearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Dettagli della transazione
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Transazione non trovata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /transactions/{transactionId}/cancel:
    put:
      tags:
        - Transactions
      summary: Cancella transazione
      description: Cancella una transazione in stato PENDING
      security:
        - bearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Transazione cancellata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Errore nella cancellazione
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                cannot_cancel:
                  summary: Transazione non cancellabile
                  value:
                    success: false
                    message: "Impossibile annullare: Solo transazioni PENDING possono essere cancellate"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT ottenuto tramite login o registrazione.
        Durata: 24 ore per access token, 7 giorni per refresh token.
  schemas:
...
    TransactionRequest:
      type: object
      required:
        - accountId
        - amount
      properties:
        accountId:
          type: integer
          format: int64
          description: ID del conto per la transazione
        amount:
          type: string
          description: Importo della transazione (formato stringa per precisione)
          example: "150.75"
        description:
          type: string
          maxLength: 255
          description: Descrizione opzionale della transazione
    TransferRequest:
      type: object
      required:
        - fromAccountId
        - toAccountId
        - amount
      properties:
        fromAccountId:
          type: integer
          format: int64
          description: ID del conto di origine
        toAccountId:
          type: integer
          format: int64
          description: ID del conto di destinazione
        amount:
          type: string
          description: Importo del trasferimento
          example: "500.00"
        description:
          type: string
          maxLength: 255
          description: Descrizione opzionale del trasferimento

    # Response DTOs
...
    TransactionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Transazione completata con successo"
        transaction:
          $ref: '#/components/schemas/Transaction'
    PaginatedTransactionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        accountId:
          type: integer
          format: int64
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        currentPage:
          type: integer
          description: Pagina corrente (0-based)
        totalPages:
          type: integer
          description: Numero totale di pagine
        totalElements:
          type: integer
          format: int64
          description: Numero totale di elementi
        hasNext:
          type: boolean
          description: true se esiste una pagina successiva
        hasPrevious:
          type: boolean
          description: true se esiste una pagina precedente
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operazione completata con successo"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Messaggio di errore"

    # Entity models
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID univoco della transazione
        transactionId:
          type: string
          description: ID pubblico della transazione (UUID)
          example: "123e4567-e89b-12d3-a456-426614174000"
        amount:
          type: string
          description: Importo della transazione
          example: "150.00"
        description:
          type: string
          maxLength: 255
          description: Descrizione della transazione
        type:
          $ref: '#/components/schemas/TransactionType'
        status:
          $ref: '#/components/schemas/TransactionStatus'
        timestamp:
          type: string
          format: date-time
          description: Data e ora della transazione
        fromAccount:
          type: object
          description: Informazioni del conto di origine
          nullable: true
          properties:
            id:
              type: integer
              format: int64
            accountNumber:
              type: string
            ownerName:
              type: string
        toAccount:
          type: object
          description: Informazioni del conto di destinazione
          nullable: true
          properties:
            id:
              type: integer
              format: int64
            accountNumber:
              type: string
            ownerName:
              type: string

    # Enums
...
    TransactionType:
      type: string
      enum:
        - TRANSFER
        - DEPOSIT
        - WITHDRAWAL
        - PAYMENT
      description: |
        Tipologie di transazione:
        * `TRANSFER` - Trasferimento tra conti interni
        * `DEPOSIT` - Deposito (da esterno)
        * `WITHDRAWAL` - Prelievo (verso esterno)
        * `PAYMENT` - Pagamento
...
security:
  - bearerAuth: []

# Esempi di utilizzo
...